Remove depending relations model when delete parent model.

+++ tfl/repository/UnitRepository.php
@@ -2,6 +2,7 @@
 
 namespace tfl\repository;
 
+use tfl\units\UnitActive;
 use tfl\units\UnitOption;
 
 trait UnitRepository
@@ -62,6 +63,10 @@ trait UnitRepository
             return false;
         }
 
+        if (!$this->deleteRelationModels()) {
+            return false;
+        }
+
         if (!$this->deleteModel()) {
             return false;
         }
@@ -69,12 +74,63 @@ trait UnitRepository
         if (!$this->deleteModelUnit()) {
             return false;
         }
+        $this->afterDelete();
 
-        if (!$this->deleteModelRelations()) {
-            return false;
+        return true;
+    }
+
+    /**
+     * Удаление только зависимых моделей
+     * @return bool
+     */
+    private function deleteRelationModels()
+    {
+        $hasErrors = false;
+
+        foreach ($this->getUnitData()['relations'] as $attr => $data) {
+            if (!$this->hasAttribute($attr)) {
+                continue;
+            }
+
+            if ($data['link'] == UnitActive::LINK_HAS_ONE_TO_ONE) {
+                /**
+                 * @var UnitActive $model
+                 */
+                $model = $this->$attr;
+                if (!$this->deleteRelationModel($model)) {
+                    $this->addDeleteError($attr, $model->getDeleteErrors());
+                    $hasErrors = true;
+                    break;
+                }
+            } else if ($data['link'] == UnitActive::LINK_HAS_ONE_TO_MANY) {
+                foreach ($this->$attr as $index => $model) {
+                    /**
+                     * @var UnitActive $model
+                     */
+                    if (!$this->deleteRelationModel($model)) {
+                        $this->addDeleteError($attr, $model->getDeleteErrors());
+                        $hasErrors = true;
+                        break;
+                    }
+                }
+            }
         }
 
-        $this->afterDelete();
+        return !$hasErrors;
+    }
+
+    /**
+     * Удаляем дочернюю модель, если она не удаляется, выводим ошибку
+     * @param UnitActive $model
+     * @return bool
+     */
+    private function deleteRelationModel(UnitActive $model)
+    {
+        if ($model->isDependModel()) {
+            if (!$model->delete()) {
+                return false;
+            }
+        }
 
         return true;
     }
